Transpile application.js and move it to public assets
This allows us to tranform our es6 code to es5

An important thing to note about browserify is that by default
browserify takes an array of file paths and returns a text stream.
However Gulp requires each transformation to provide a virtual file as
input.

We can solve this problem by using the vinyl-source-stream package.
vinyl-source-stream is a stream transformation which takes text stream
as an input and returns a virtual file as an output.