When person fields are undefined set to ''
Otherwise the following warnings will be present
- Warning: ... is changing an uncontrolled
input of type text to be controlled. Input elements should not switch
from uncontrolled to controlled (or vice versa). Decide between using a
controlled or uncontrolled input element for the lifetime of the
component

- Warning: `value` prop on `input` should not be null.
Consider using the empty string to clear the component or `undefined`
for uncontrolled components.