Capybara configuration changes
Support a known time zone that is not UTC for tests to prove date
conversion to local time zone.

Support ignoring accessiblity checks for specific elements. Added as
date picker component adds aria tags that are temporarily invalid, until
the component is fully rendered after user interaction. At the time of
this commit we are going to have the aria tags tested manually on a
machine with the accessiblity software

[#145765803]