Fix copy button spec for parallel specs
For some reason this test was failing when run in parallel on jenkins
but not when run in parallel locally. We suspected the native send_keys
to be causing the issue due to it possibly not supporting multiple
threading correctly for the <ctrl>-v we were using. Once changed to just
send_keys it worked.

[#148018893]