saveContactSaga updates/creates contact based on payload
saveContactSaga calls update contact API when contact id is present
and calls create contact API when contact id is not present.

Adding this complexity to saga instead of the view allows the view to
remain ignorant to the difference between saving and creating a contact
since visually they are no different.

[#151030845]