Call SecurityRepository.retrieve_security_token in controller
This will get the token from perry for the corresponding access code if
the flag perry_version_two is enabled. Otherwise it will return the
token that is passed in. Version two passes the access code as
accessCode on the query params as apposed to version one that passes
token on the query params. So for now the method takes both so the
controller doesn't have to know about the feature flag.

[INT-459]