Update Jenkinsfile to find PR branch names
Jenkins gives different environment variables for finding the git
branch name depending on whether the job is configured for for all
GitHub branches or only Pull Requests. We use the branch name to
determine some publish and deploy behavior for the master branch,
so this prevents falsely detecting PRs as master.