Check autocomplete_search_bar results for name highlights
This commit passes the previously included tests, to ensure that
we check autocomplete fields for highlighted names. The end result
is that searching for a partial name (eg "Alex") will still
highlight "Alexander" as it did with the previous ElasticSearch
index. We still defer to an exactly matched name if present, and
fallback to the non-emphasized name if no match is found.